Moved relevant classes from messages to auctions package.
