+++ src/functional/java/org/economicsl/auctions/ImperativePeriodicDoubleAuction.java
@@ -1,8 +1,8 @@
 package org.economicsl.auctions;
 
 
-import org.economicsl.auctions.messages.Accepted;
-import org.economicsl.auctions.messages.Rejected;
+import org.economicsl.auctions.Accepted;
+import org.economicsl.auctions.Rejected;
 import org.economicsl.auctions.singleunit.OpenBidAuction;
 import org.economicsl.auctions.singleunit.orders.SingleUnitOrder;
 import org.economicsl.auctions.singleunit.pricing.MidPointQuotePricingPolicy;
