+++ src/main/java/org/economicsl/auctions/singleunit/InsertResult.java
@@ -16,8 +16,8 @@ limitations under the License.
 package org.economicsl.auctions.singleunit;
 
 
-import org.economicsl.auctions.messages.Accepted;
-import org.economicsl.auctions.messages.Rejected;
+import org.economicsl.auctions.Accepted;
+import org.economicsl.auctions.Rejected;
 import scala.util.Either;
