+++ src/main/scala/org/economicsl/auctions/actors/AuctionParticipantActor.scala
@@ -16,8 +16,7 @@ limitations under the License.
 package org.economicsl.auctions.actors
 
 import akka.actor.ActorRef
-import org.economicsl.auctions.{AuctionParticipant, AuctionProtocol}
-import org.economicsl.auctions.messages.{Accepted, Canceled, Rejected}
+import org.economicsl.auctions._
 import org.economicsl.core.Tradable
 import org.economicsl.core.util.Timestamper
